devhub-1025 Adding dynamic folder structure to support dynamic l1 and l2 url structuring
